-update:
now there are no more backend calls (to icmm service) in the directive
controller of the pre-triage chart widget. Requirement therefore is that
the
backend call in debriefing view controller returns the full data (which
can be achieved by setting the level parameter in the get-request (i.e.
to the value 5)).