[issue #7] removed backend calls in directive controller:
now there are no more backend calls (to icmm service) in the directive
controller of the care measures chart widget. Requirement therefore is
that the
backend call in debriefing view controller returns the full data (which
can be achieved by setting the level parameter in the get-request (i.e.
to the value 5)).

-switched dataSource to a development instance of the icmm service in
order to have proper test data